requirements.txt
aiohttp==3.7.3
amqp>=2.5.2
appdirs==1.4.4
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
celery>=4.4.2
certifi==2020.12.5
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
distlib==0.3.1
Django==3.1.5
boto3>=1.13.16
botocore>=1.16.16
django-celery-results==2.0.1
django-friendship==1.9.1
django-lrucache-backend==4.0.0
django-push-notifications>=2.0.0
django-rest-framework==0.1.0
django-slowtests==1.1.1
django-timezone-field==4.1.1
djangorestframework==3.12.2
filelock==3.0.12
gunicorn==20.0.4
identify==1.5.13
idna==2.10
IMDbPY==2020.9.25
importlib-metadata==3.4.0
jikanpy==4.2.2
kombu>=4.6.8
mal-api==0.3.7
multidict==5.1.0
nodeenv==1.5.0
pre-commit==2.9.3
prompt-toolkit==3.0.11
psycopg2-binary==2.8.6
PyJWT==1.4.0
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.5
redis==3.5.2
regex==2020.11.13
requests==2.25.1
simplejson==3.17.2
six==1.15.0
soupsieve==2.1
SQLAlchemy==1.3.22
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
urllib3>=1.21.1
vine>=1.3.0
virtualenv==20.4.0
wcwidth==0.2.5
yarl==1.6.3
zipp==3.4.0
tmdbsimple==2.2.15

